requirements.txt
discord.py
python-dotenv
pyodbc
ollama